+++ Sources/耗材管家/AppNotifications.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension Notification.Name {
+    static let filamentDataChanged = Notification.Name("filamentDataChanged")
+}